Fix grammar test

+++ projects/batfish-common-protocol/src/test/java/org/batfish/datamodel/matchers/IsisInterfaceLevelSettingsMatchers.java
@@ -21,8 +21,7 @@ public final class IsisInterfaceLevelSettingsMatchers {
    * Provides a matcher that matches when the {@link IsisInterfaceLevelSettings}'s cost is {@code
    * expectedCost}.
    */
-  public static @Nonnull Matcher<IsisInterfaceLevelSettings> hasCost(
-      @Nullable Integer expectedCost) {
+  public static @Nonnull Matcher<IsisInterfaceLevelSettings> hasCost(@Nullable Long expectedCost) {
     return new HasCost(equalTo(expectedCost));
   }
